+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3285,6 +3285,52 @@ typedef union rocprofiler_hip_api_args_t
         unsigned long long* streamId;
     } hipStreamGetId;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 15
+    struct
+    {
+        hipLibrary_t*     library;
+        const void*       code;
+        hipJitOption*     jitOptions;
+        void**            jitOptionsValues;
+        unsigned int      numJitOptions;
+        hipLibraryOption* libraryOptions;
+        void**            libraryOptionValues;
+        unsigned int      numLibraryOptions;
+    } hipLibraryLoadData;
+    struct
+    {
+        hipLibrary_t*     library;
+        const char*       fileName;
+        hipJitOption*     jitOptions;
+        void**            jitOptionsValues;
+        unsigned int      numJitOptions;
+        hipLibraryOption* libraryOptions;
+        void**            libraryOptionValues;
+        unsigned int      numLibraryOptions;
+    } hipLibraryLoadFromFile;
+    struct
+    {
+        hipLibrary_t library;
+    } hipLibraryUnload;
+    struct
+    {
+        hipKernel_t* pKernel;
+        hipLibrary_t library;
+        const char*  name;
+    } hipLibraryGetKernel;
+    struct
+    {
+        unsigned int* count;
+        hipLibrary_t  library;
+    } hipLibraryGetKernelCount;
+#endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 16
+    struct
+    {
+        hipStream_t dst;
+        hipStream_t src;
+    } hipStreamCopyAttributes;
+#endif
 } rocprofiler_hip_api_args_t;
 
 ROCPROFILER_EXTERN_C_FINI